requirements.txt
annotated-types==0.7.0
archinfo==9.2.193
attrs==25.4.0
autocommand==2.2.2
Automat==25.4.16
bitarray==3.8.0
bitstring==4.3.1
borgbackup==1.4.3
btrfsutil==6.19
build==1.4.0
capstone==5.0.7
certifi==2026.1.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
constantly==23.10.4
cryptography==46.0.5
CT3==3.4.0
cupshelpers==1.0
dbus-python==1.4.0
defusedxml==0.7.1
distro==1.9.0
dnspython==2.8.0
filebytes==0.10.2
filelock==3.24.0
gitdb==4.0.12
GitPython==3.1.46
Glances==4.5.0.5
gpg==2.0.0b25
httpie==3.2.4
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
iniconfig==2.1.0
installer==0.7.0
jaraco.collections==5.1.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaraco.text==4.0.0
jeepney==0.9.0
Jinja2==3.1.6
legacy-cgi==2.6.4
lensfun==0.3.4
libmsym==0.2.4
lit==21.1.8.dev0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.8.0
msgpack==1.1.2
multidict==6.7.1
natsort==8.4.0
netsnmp-python==1.0a1
nftables==0.1
OpenTimelineIO==0.18.1
orjson==3.11.7
outcome==1.3.0.post0
packaging==26.0
peewee==3.19.0
pillow==12.1.1
pkg_resources==81.0.0
platformdirs==4.9.3
pluggy==1.6.0
psutil==7.2.2
pyaaf2==1.7.1
pycairo==1.29.0
pycparser==2.23
pycups==2.0.4
pydantic==2.12.5
pydantic_core==2.41.5
pyfuse3==3.4.2
Pygments==2.19.2
PyGObject==3.54.5
pymongo==4.16.0
pyproject_hooks==1.2.0
PyQt5==5.15.11
PyQt5_sip==12.18.0
PyQt6==6.10.2
PyQt6_sip==13.11.0
PySide6==6.10.2
PySocks==1.7.1
pytest==8.4.2
python-dateutil==2.9.0
pyvex==9.2.197
PyYAML==6.0.3
QTermWidget==2.3.0
QtPy==2.4.3
Reflector==2023.6.28.0.36.1
regex==2026.2.19
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
ropper==1.13.13
SecretStorage==3.3.3
setuptools==82.0.0
shiboken6==6.10.2
shiboken6-generator==6.10.2
shtab==1.8.0
six==1.17.0
smbus==1.1
smmap==6.0.0
sniffio==1.3.1
sortedcontainers==2.4.0
spglib==2.7.1.dev0+g12355c77f.d20260111
TBB==0.2
torbrowser-launcher==0.3.7
trio==0.33.0
Twisted==24.7.0
typing-inspection==0.4.2
typing_extensions==4.15.0
ufw==0.36.2
urllib3==2.6.3
vorta==0.11.3
wheel==0.46.3
yt-dlp==2026.2.21
yt-dlp-ejs==0.5.0
ytmusicapi==1.11.5
z3-solver==4.15.4.0
zope.interface==8.2
